+++ src/Event/TwigCollectBlocksEvent.php
@@ -25,26 +25,21 @@
 
 declare(strict_types=1);
 
-namespace Machinateur\TwigBlockValidator\Event\Validator;
+namespace Machinateur\TwigBlockValidator\Event;
 
-use Machinateur\TwigBlockValidator\Event\NotifiableInterface;
-use Machinateur\TwigBlockValidator\Event\NotifiableTrait;
-use Symfony\Component\Finder\Finder;
+use Machinateur\TwigBlockValidator\Twig\Node\TwigBlockStackInterface;
 
-class TwigLoadFilesEvent implements NotifiableInterface
+/**
+ * @phpstan-import-type _Block              from TwigBlockStackInterface
+ */
+readonly class TwigCollectBlocksEvent
 {
-    use NotifiableTrait;
-
-    public const CALL_BEGIN = 'begin';
-    public const CALL_END   = 'end';
-    public const CALL_STEP  = 'step';
-
     /**
      * @param array<string> $paths
+     * @param array<_Block> $blocks
      */
     public function __construct(
-        public readonly string $namespace,
-        public readonly array  $paths,
-        public readonly Finder $finder,
+        public array $paths,
+        public array $blocks,
     ) {}
 }
