+++ src/Validator/TwigBlockValidator.php
@@ -29,12 +29,6 @@ namespace Machinateur\TwigBlockValidator\Validator;
 
 use Composer\Semver\Semver;
 use Machinateur\TwigBlockValidator\Event\Validator\ValidateCommentsErrorEvent;
-use Machinateur\TwigBlockValidator\Event\Validator\TwigLoadFilesEvent;
-use Machinateur\TwigBlockValidator\Event\Validator\TwigLoadPathsErrorEvent;
-use Machinateur\TwigBlockValidator\Event\Validator\TwigLoadPathsEvent;
-use Machinateur\TwigBlockValidator\Event\Validator\TwigRegisterPathsErrorEvent;
-use Machinateur\TwigBlockValidator\Event\Validator\TwigRegisterPathsEvent;
-use Machinateur\TwigBlockValidator\Event\Validator\TwigCollectBlocksEvent;
 use Machinateur\TwigBlockValidator\Event\Validator\ValidateCommentsEvent;
 use Machinateur\TwigBlockValidator\Service\NamespacedPathnameBuilder;
 use Machinateur\TwigBlockValidator\Service\TwigBlockResolver;
@@ -43,8 +37,6 @@ use Machinateur\TwigBlockValidator\Twig\Extension\BlockVersionExtension;
 use Machinateur\TwigBlockValidator\Twig\Node\CommentCollectionInterface;
 use Machinateur\TwigBlockValidator\Twig\Node\TwigBlockStackInterface;
 use Psr\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\Finder\Finder;
-use Symfony\Component\Finder\SplFileInfo;
 use Twig\Error\Error as TwigError;
 use Twig\Error\LoaderError;
 use Twig\Error\RuntimeError;
